refactor: AccountAdapter: set up an OnViewClick interface to handle the clicking of views and pass actions back to the implementing class. remove member LayoutInflater add functions to add, update and delete accounts from the account list and notify the data changed. move sorting into the adapter move calculating total accounts balance into the adapter
Checkbook: keep static instance of Checkbook.
use longs for account ids

CheckbookDbHelper: use longs for the account ids.

AccountsFragment: move intent keys to seperate class.
refactor handling the passing of data between the adapter and the fragment using the new adapter OnViewClick inteferface.
move the refreshing of the total balance view to function.
only remove the account from the adapter if we were sucessful in removing the account from the database.
handle adapter data updates in the results from activities opened from within the fragment.
handle launching an account activity for the selected account in the newly created OnViewClickListener interface of the adapter.

AccountIntentKeys:
add class to define the various intent arg keys used by the account classes.

Account: make account implement Comparable, and also Serializable so we can sort the lists of accounts, and
add an account to an intent bundle.
Use longs for the account ids.

ActivityRequests: create class to define the ActivityRequest ints used when launching Activities.

Fragment Requests: create a class to define the request ints when launching Fragments.

MoneyTextWatcher: add credit from where I got most of the code froma nd how i edited the code to work for my application.

AccountActivity: Switch to using the static Checkbook instance.
Send an intent back from the activity result so we can pass the data (account position) back to the class that launched the activity.
Account ids are longs now.

AddEditAccountActivity: account ids are longs now.
Use static instance of Checkbook
Refactor logic to create an account object when we create a new account, and we send this account back to the launching activity/ragment as a serialized intent object
so we can update that account in the adapter acocunt list.
We can still use a bool for updating, but we need to pass the updated object and the position back in the intent.
We can get the account id from the database insert function and use it to construct an account object.
Mkae sure we update the instance account details when we update them so the data gets passed back to the adapter list.

MainActivity: we only need to initialize the checkbookfrom the MainActivity now.
We move the FragmentRequests out to their own class.