Change intent keys, request codes and result codes to interfaces.
AccountAdapter: store only the applicationContext() reference to avoid potential memory leaks.
Infalte view using the parents context since we are only storing the application context.

TransactionAdapter: this code will be changed drastically to a recyclerview adapter eventually, but for now, switch to using the public database date formats exposed in the Checkbook helper class.
Also rename textview for name to payee to better reflect its contents. Use the database function to get the payee name given the payee id. Only store the application context.

<All Fragments>: Remove the horrible memory leaks of storing static references to each fragment created. What was I thinking???!!!??? Big no no.

AccountsFragment: use new Checkbook.getInstance(Context context) to obtain data from the database.

CheckbookDbHelper -> Checkbook: remove the goofy in-between class of Checkbook. Use a singleton pattern for the Checkbook database helper class. Store static reference to the database helper and return the static ref when accessed after it has initially been set
up. Also make sure to pass in the application context via getApplicationContext() on the incoming context because we don't know where the context could be coming from and we want to make sure to initialize it with the application context.
Only set foreign key support in onOpen for pre Jellybean devices and onConfigure for Jellybean and later devices.
Use public Strings for the date format strings so we remove the getters.
Create a dummy method for getPayeeName()

Account: remove the setScale when returning the balances from the Accounts because we set the scale before adding to the UI.

Transaction: change id fields to longs and also remove the setScale from the amount getter.
Fill in parceling logic.

Transfer: switch id fields to long.
Implement parcelable and fill in the logic.

MoneyTextWatcher: remove chance for memory leak by only storing application context.
Fix the checking if the amount entered is zero so it can be set to an empy string logic, as I broke it when I removed the currency symbol. Get the fractional digits of the locale currency when settings the scale of the amount.
Also get the number of digits once and store locally.

AccountActivity: get the coordinator layout and make sure it isn't null before attempting to show the snackbar.
Use the newly rated Checkbook singleton pattern to access database.

AddEditAccount: use the newly created singleton pattern to access the database.

MainActivity: we know longer need to create a database instance in the mainactivity. We will just create it the first time we need it.

transaction_listview_layout: rename the transaction_name textview to transaction_payee.