Activity Layouts: remove all AppBars. We are going elevation-less until i can figure out the bug where the shadow disapears when a snackbar is shown.
fragment_accounts: use no elevation on our AppBarLayout wrapper for the total balance view.
use the attr primary color instead of explicitly calling the color resource.

AccountsFragment: keep an instance reference to the account list, so we are not constantly requerying the database for the accounts. This is what the static instanceof the fragment was for before, but I didn't know as much about context leaks and why it is a bad idead to
save a static reference to a fragment, so this is a better solution. The view still gets recreated, however we bundle out and save imprtant aspects of the view.
Convert the total balance BigDecimal to a local variable.
Bundle the two refresh methods into one that we call anytime thier is a potential change in the account list.