AccountAdapters: call size() on account list once.
TransactionAdapter: switch over to recyclerview adapter and get some of the basics filled in. Still lots more to do here.

CheckbookApplication: extend my own application in order to use LeakCanary.

AccountFragment: replace AccountActivity with AccountFragment. Eventually I will replace all activities with fragments and just use the MainActivity as the maintainer for the fragments.

AccountsFragment -> AccountsListFragment: define navigation callback interface so we can pass of the marshalling of launching new fragments to the MainActivity.
Set the title in onResume since we cann add fragments to the back stack from this fragment.
Use navigation callbacks to launch the AccountFragment(in place of the AccountActivity).

NavigationDrawerFragment: create a public method to lock/unlock the drawer and hide/show the indicator for use when we launch a fragment on the backstack from another fragment. This allows us to implement back navigation like in a normal activity.

Checkbook: create a dummy method to create a empty ArrayList of Transactions.

MainActivity: implement the AccountListNav callbacks and the AccountNav callbacks.
Add a back stack listener to handle unlocking the drawer once we clear the backstack.
Add logic in onBackPressed() to check the backstack. If there is entries in the backstack, pop them off. Otherwise get the account list fragment and make sure it is displayed
Implement the launchAccountFragment callback to lock the drawer and add the AccountFragment to the backstack.