AccountFragment: create OnAccountUpdateCallbacks interface to delegate the changed account data to the MainActivity, which then forwards it to the AccountListFragment. Get the current account name from the mAccount instance if it is not null. Check to make sure mLayoutManager is not null before we attempt to parcelize it as well as mTransactionList. Create finish() method to simulate a backpress to return back from Fragment, Use navigation callback to launch and AddEditAccountFragment instance from the MainActivity. Make sure we inform the OnAccountUpdatedCallbacks that an account was deleted.
AccountListFragment: use the navigation callback to launch a new AddEditAccountFragment instace on FAB button click.
Use instanceof on context to ensure activity implements the callbacks.
Check that the mLayoutManager and mAccountList are not null before trying to parcel them.
Use navigation callback to launch the AddEditAccount when the longpress edit menu is selected.
Remove the old activity stuff as we are no longer using activities outside of the MainActivity.
Create public API calls to handle the onAccountAdded, onAccountDeleted and onAccountUpdated calls from the MainActivity.

AddEditAccountActivity -> AddEditAccountFragment: implement the same functionaility using a fragment instead of the activity. Still needs work as before in implementing the transactions and transfers.

SettingsFragment: the RuntimeException message used an outdated interface name.

Remove the interfaces used to store the request codes, results codes and intent argument keys.

MainActivity: use a private static Handler class and a static final anonymous Runnable class to prevent context and memory leaks when we post a
new runnable to implement the double tap to exit functionaility.
Add transition animations to all the fragment transactions.
Change instance mDoubleBackToExitPressedOnce boolean to a static member.
Add in interface methods and fill out the logic to launch  the AddEditAccountFragment, delete, upate and add accounts.

fragment_account: fix derp in using the margin (16dp) instead of the elevation (4dp) for the fab. It looked awful lol.