AccountFragment: set up onClickListener for the add transaction FAB to launch an AddEditTransactionFragment through the navigation callbacks. Add public API functions to handle transaction modofication once we implement everything, Add interface method to launch an AddEditTransactionFragment from the host activity.
AddEditTransactionFragment: set up the basic shell for the AddEditTransactionFragment.

AddEditTransferFragment: not sure if I am going to use this or not yet because I haven't decided how I am going to implement handling transfers as opposed to transaction. Empty fragment for now.

NavigationDrawerFragment: remove the drawer shadow because it is handled in the new deisgn library implementations.
Switch to checking the context for instanceof of the interface instead of blindly trying to cast the activity to the interface.

Delete AddEditTransaction and AddEditTransfer activities. Nuke all the activities!!

MainActivity: close nav drawer first and return if it is open before checking on the double back to exit.
Implement the launchAddEditTransaction callback to launch a new AddEditTransactionFragment.
Override the AddEditTransaction OnTransactionUpdated interface methods.

handle.xml -> empty_shape.xml

Remove drawer shadow .9 png

fragment_add_edit_transaction: start working on the ui layout for the AddEditTransaction page.

fragment_add_edit_transfer: empty fragment for now. Not sure if i will even use this fragment.

menu_add_edit_transaction: create a menu for the AddEditTransactionFragment.

strings.xml: add hint strings for the transaction Payee and amount text inputs.